// src/types/auth.types.ts
export interface RegisterDTO {
    email: string;
    password: string;
    name?: string;
}

export interface LoginDTO {
    email: string;
    password: string;
}

export interface AuthResponse {
    token: string;
    user: {
        id: number;
        email: string;
        name?: string;
    };
}